corrected smr formula

+++ pojo-ops/src/main/java/com/stanzaliving/core/operations/dto/report/food/summary/DateLevelNumbersDto.java
@@ -32,7 +32,7 @@ public class DateLevelNumbersDto {
 
 	private int movedInResidents;
 	
-	private int activeMeals;
+	private double activeMeals;
 
 	private String residenceUuid;
